Backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
SQLAlchemy==2.0.29
